+++ mido/messages/messages.py
@@ -11,6 +11,7 @@ PY2 = (sys.version_info.major == 2)
 class BaseMessage(object):
     """Abstract base class for messages."""
     is_frozen = False
+    is_meta = False
 
     def copy(self):
         raise NotImplemented